requirements.txt
```text
# --- UI -------------------------------------------------
streamlit>=1.34

# --- LangChain + providers ------------------------------
langchain>=0.2
langchain-community>=0.2          
langchain-openai>=0.1
langchain-mistralai>=0.1
openai>=1.30

# --- Embeddings / vector store --------------------------
faiss-cpu>=1.8

# --- NLP / tokenisation ---------------------------------
tiktoken>=0.6

# --- Document parsing -----------------------------------
PyMuPDF>=1.24          # import name: fitz
python-docx>=1.1
beautifulsoup4>=4.12
markdown>=3.5
pandas>=2.2
openpyxl>=3.1
xlrd>=2.0
pytesseract>=0.3
Pillow>=10.3

# --- Utilities ------------------------------------------
python-dotenv>=1.0
pydantic>=2.7
